GOLD-TIER/ai_employee/pyproject.toml
[project]
name = "ai-employee-gold"
version = "3.0.0"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "schedule>=1.2.0",

    # Silver Tier — NEW additions:
    "linkedin-api>=2.0.0",
    "requests>=2.31.0",
    "python-crontab>=3.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",

    # Gold Tier — NEW additions:
    "tweepy>=4.14.0",
    "facebook-sdk>=3.1.0",
    "instagrapi>=2.0.0",
    "xmlrpc-client>=0.0.1",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
    "orjson>=3.9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
